add validation in web and update category&product

+++ src/MyShop.Application/Commands/CreateProduct.cs
@@ -2,4 +2,4 @@
 
 namespace MyShop.Application.Commands;
 
-public sealed record CreateProduct(string Name, string Description, double Price, Guid CategoryId) : ICommand;
+public sealed record CreateProduct(string Name, string Description, decimal Price, Guid CategoryId) : ICommand;
